Refactored entities into separate packages for separation
