+++ backend/src/main/java/com/speakfluid/backend/entities/json/JSONLog.java
@@ -1,4 +1,4 @@
-package com.speakfluid.backend.entities;
+package com.speakfluid.backend.entities.json;
 
 import java.util.Map;
