+++ backend/src/main/java/com/speakfluid/backend/entities/steps/CaptureStep.java
@@ -1,7 +1,9 @@
-package com.speakfluid.backend.entities;
+package com.speakfluid.backend.entities.steps;
+
+import com.speakfluid.backend.entities.message.Dialogue;
+import com.speakfluid.backend.entities.message.Message;
 
 import java.lang.String;
-import java.lang.reflect.Array;
 import java.util.*;
 import static java.util.Map.entry;
