+++ backend/src/main/java/com/speakfluid/backend/entities/steps/ImageStep.java
@@ -1,4 +1,8 @@
-package com.speakfluid.backend.entities;
+package com.speakfluid.backend.entities.steps;
+
+import com.speakfluid.backend.entities.message.Dialogue;
+import com.speakfluid.backend.entities.message.Message;
+import com.speakfluid.backend.entities.steps.TalkStep;
 
 import java.util.Arrays;
 import java.util.List;
