+++ backend/src/main/java/com/speakfluid/backend/entities/steps/TextStep.java
@@ -1,4 +1,7 @@
-package com.speakfluid.backend.entities;
+package com.speakfluid.backend.entities.steps;
+
+import com.speakfluid.backend.entities.message.Dialogue;
+import com.speakfluid.backend.entities.message.Message;
 
 import java.util.Arrays;
 import java.util.List;
